refactor member auth
changes memberverify to Trelora::REST::Client
adds method
-authenticate
-token
changes sessions controller
